% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtMergingLOOCV.R
\name{cdtMergingLOOCV}
\alias{cdtMergingLOOCV}
\title{CDT Merging Leave-One-Out Cross-Validation}
\usage{
cdtMergingLOOCV(time.step = "dekadal", start.date = c(1981, 1, 1),
  end.date = c(2018, 12, 3), station = list(file = NULL, sep = ",",
  miss = "-99"), netcdf = list(dir = NULL, varid = NA, order.lon = 1,
  order.lat = 2, format = "rfe_\%s\%s\%s.nc"), merging.method = "SBA",
  interp.method = "idw", spheric = FALSE, maxdist = 2.5,
  pass.ratio = c(1, 0.5, 0.2), pass.nmin = c(5, 4, 3),
  pass.nmax = c(15, 10, 7), neg.value = FALSE, output.file = NULL,
  use.RnoR = FALSE, pars.RnoR = list(wet.day = 1, smooth = FALSE),
  vgm.model = c("Exp", "Gau", "Sph", "Pen"), parallel = list(dopar =
  TRUE, detect.cores = TRUE, nb.cores = 2))
}
\arguments{
\item{time.step}{Time step of the data. Should be \strong{"daily"}, \strong{"pentad"}, \strong{"dekadal"} or \strong{"monthly"}.}

\item{start.date}{A vector of the start date to merge. 
\itemize{
    \item \emph{Daily data}: \code{c(year, month, day)}
    \item \emph{Pentad data}: \code{c(year, month, pentad)},  \code{pentad} must be 1, 2, 3, 4, 5, or 6
    \item \emph{Dekadal data}: \code{c(year, month, dekad)},  \code{dekad} must be 1, 2 or 3
    \item \emph{Monthly data}: \code{c(year, month)}
}}

\item{end.date}{A vector of the start date to merge. Same format as \code{start.date}.}

\item{station}{A list containing the station information in the form \code{list(file, sep, miss)}.
\itemize{
  \item \code{file}: full path to the station file (CDT station data format)
  \item \code{sep}: separator of the column
  \item \code{miss}: missing value flag
}}

\item{netcdf}{A list containing the netcdf data information in the form \code{list(dir, order.lon, order.lat, format)}.
\itemize{
 \item \code{dir}: full path to the directory containing the ne netcdf data
 \item \code{varid}: name of the variable to be used
 \item \code{order.lon}: order of the longitude (netcdf dimension order)
 \item \code{order.lat}: order of the latitude
 \item \code{format}: filename format of the netcdf files
}}

\item{merging.method}{Merging method to be used. Valid options: \strong{"CSc"}, \strong{"BSc"}, \strong{"SBA"} or \strong{"RK"}.
\itemize{
 \item \strong{"CSc"}: Cressman Scheme
 \item \strong{"BSc"}: Barnes Scheme
 \item \strong{"SBA"}: Simple Bias Adjustment
 \item \strong{"RK"}: Regression Kriging
}}

\item{interp.method}{Interpolation method to be used for \strong{"SBA"} and \strong{"RK"}.
 Valid options: \strong{"idw"}, \strong{"shepard"}, \strong{"spheremap"} or \strong{"kriging"}.
\itemize{
 \item \strong{"idw"}: Inverse distance weighted
 \item \strong{"shepard"}: Modified Shepard interpolation
 \item \strong{"spheremap"}: Spheremap interpolation method
 \item \strong{"kriging"}: Ordiranry kriging
}}

\item{spheric}{If \code{FALSE} (default), then a cartesian distance will be computed. If set to \code{TRUE}, a spherical distance (using a standard great circle method) will be computed.}

\item{maxdist}{Maximum radius of influence in decimal degree.}

\item{pass.ratio}{A vector giving the fraction of \code{maxdist} to be used for each pass.}

\item{pass.nmin}{A vector giving the minimum number of stations to be used to interpolate a grid point for each pass. Must be the same length as \code{pass.ratio}.}

\item{pass.nmax}{A vector giving the maximum number of stations to be used to interpolate a grid point for each pass. Must be the same length as \code{pass.ratio}.}

\item{neg.value}{If \code{TRUE}, negative values will be kept. If \code{FALSE} (default), negative values will be set to zero.}

\item{output.file}{Full path name to save the result.}

\item{use.RnoR}{If \code{TRUE}, apply rain-no-rain mask for rainfall data.}

\item{pars.RnoR}{A list of parameters to be used if \code{use.RnoR} is \code{TRUE}.
\itemize{
 \item \code{wet.day}: wet day definition
 \item \code{smooth}: if \code{TRUE}, smooth the mask
}}

\item{vgm.model}{A vector of variogram model to be used if \code{interp.method} is \strong{"kriging"}. Default is \code{c("Exp", "Gau", "Sph", "Pen")}.}

\item{parallel}{A list of parameters for parallel processing.
\itemize{
\item \code{dopar}: if \code{TRUE} (default), use parallel computing
\item \code{detect.cores}: if \code{TRUE} (default), detect the number of CPU cores.
\item \code{nb.cores}: if \code{dopar} is \code{TRUE} and \code{detect.cores} is \code{FALSE}, the number of CPU cores to be used 
}}
}
\description{
This function perform a Leave-One-Out Cross-Validation.
}
\examples{

\dontrun{
cdtMergingLOOCV(
     time.step = "dekadal",
     start.date = c(2018, 1, 1),
     end.date = c(2018, 12, 3),
     station = list(file = "~/DONNEES/STN_DATA/CDT_RR_dekad_1981-2018_CLM.csv",
                 sep = ",", miss = "-99"),
     netcdf = list(dir = "~/DONNEES/GRIDDED_DATA/tamsat_v3/dekadal", varid = "rfe",
                 order.lon = 1, order.lat = 2, format = "rfe_\%s\%s\%s.nc"),
     merging.method = "SBA", 
     interp.method = "idw",
     spheric = TRUE,
     maxdist = 1.0,
     pass.ratio = c(1, 0.75, 0.5),
     pass.nmin = c(5, 4, 3),
     pass.nmax = c(15, 10, 7),
     neg.value = FALSE,
     output.file = "~/DONNEES/OUT_Merging/LOOCV_SBA-idw.csv",
     use.RnoR = TRUE,
     pars.RnoR = list(wet.day = 1.0, smooth = FALSE),
     vgm.model = c("Exp", "Gau", "Sph", "Pen")
 )
}

}
