\name{grid2pointINDEX}
\alias{grid2pointINDEX}
\title{Extract grid index}
\usage{
grid2pointINDEX(pts_Coords, grd_Coords)
}
\arguments{
\item{pts_Coords}{A named list of coordinates list(lon = lon, lat = lat)}
\item{grd_Coords}{A named list of coordinates list(lon = lon, lat = lat)}
}
\value{
An integer vector.
}
\description{
Get the index of points over grid.
}
