% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtNetCDF_dataSeries.R
\name{ncFilesInfo}
\alias{ncFilesInfo}
\title{Parse the NetCDF files}
\usage{
ncFilesInfo(Tstep, start.date, end.date, months, ncDir, ncFileFormat,
  error.msg)
}
\arguments{
\item{Tstep}{The time step of the NetCDF files. Valid options: 'daily', 'pentad', 'dekadal', 'monthly'.}

\item{start.date}{An object of class \code{Date} indicating the start date to be parsed.}

\item{end.date}{An object of class \code{Date} indicating the end date to be parsed.}

\item{months}{An integer vector indicating the months to be parsed.}

\item{ncDir}{The full path to the directory containing the NetCDF files.}

\item{ncFileFormat}{Filename format of the NetCDF files. Ex: rfe_%s%s%s.nc}

\item{error.msg}{A message to be displayed when no files were found.}
}
\value{
A list of a character vector indicating the date, a character vector containing the full path name of the NetCDF files and a logical vector indicating if the file exists.
}
\description{
Parse all NetCDF data inside a directory according the provided filename format
}
