% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdtUtilities_functions.R
\name{read.CDTstation}
\alias{read.CDTstation}
\title{Read CDT station data format}
\usage{
read.CDTstation(stn.file, sep = ",", na = "-99")
}
\arguments{
\item{stn.file}{full path name to the station file (CDT station data format)}

\item{sep}{the field separator character of the column, default is comma.}

\item{na}{missing value flag, the default is "-99".}
}
\value{
A list of the id, longitude, latitude, elevation (if available) of the stations as a vector, date and the station data as a matrix of dimensions nrow= number of stations and ncol= length of date.
}
\description{
This function reads a CDT station data format.
}
